+++ modules/base/concepts/ServerRoutes.js
@@ -6,4 +6,8 @@ export default class ServerRoutes extends GeneratingConcept {
   get directory() {
     return 'server/routes'
   }
+
+  get recursive() {
+    return true
+  }
 }
