+++ modules/sitemap/SitemapRenderer.js
@@ -1,5 +1,3 @@
-import VUE_I18N_LOCALES from '#ioc/config/VUE_I18N_LOCALES'
-
 export default class SitemapRenderer {
   withStore(store) {
     this._store = store
@@ -49,7 +47,7 @@ export default class SitemapRenderer {
   }
 
   _getUrlPrefix() {
-    if (this._store.name === VUE_I18N_LOCALES[0].name) return ''
+    if (this._store.prefix === '/') return ''
     return `/${this._store.name}`
   }
 }
