+++ modules/sitemap/concepts/SitemapLocales.js
@@ -1,0 +1,9 @@
+//@ts-check
+
+import { GeneratingConcept } from '@storefront-x/core'
+
+export default class SitemapLocales extends GeneratingConcept {
+  get directory() {
+    return 'sitemap/locales'
+  }
+}
