+++ modules/vue-router-simple/sitemap/staticPages.ts
@@ -1,14 +1,12 @@
 import { routes } from '~/.sfx/pages'
 
-export default async ({ store }: any) => {
+export default async () => {
   const newRoutes = routes.slice(1, -1)
-  const storePrefix = store.prefix
 
   const urls = []
 
   for (const route of newRoutes) {
-    const routePath = route.readablePath.replace('/', '')
-    urls.push({ loc: storePrefix + routePath })
+    urls.push({ loc: route.readablePath })
   }
 
   return urls
